# src/main/java/dev/ilkersahin/java/spring/gym/controller/TrainerController.java
package dev.ilkersahin.java.spring.gym.controller;

import dev.ilkersahin.java.spring.gym.dto.request.*;
import dev.ilkersahin.java.spring.gym.dto.response.ErrorResponse;
import dev.ilkersahin.java.spring.gym.dto.response.UserCreateResponse;
import dev.ilkersahin.java.spring.gym.dto.view.TrainerWithListView;
import dev.ilkersahin.java.spring.gym.dto.view.TrainingView;
import dev.ilkersahin.java.spring.gym.exception.UserAlreadyActiveException;
import dev.ilkersahin.java.spring.gym.exception.UserAlreadyInactiveException;
import dev.ilkersahin.java.spring.gym.security.selfservice.SelfService;
import dev.ilkersahin.java.spring.gym.service.TrainerService;
import io.micrometer.core.annotation.Timed;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.format.annotation.DateTimeFormat;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.time.LocalDate;
import java.util.List;

@RestController
@RequestMapping("/api/v1/trainers")
@RequiredArgsConstructor
@Slf4j
@Tag(name = "Trainers", description = "Trainer registration and profile management")
@Timed(
        value = "gym.http.trainer",
        description = "Trainer controller HTTP requests"
)
public class TrainerController {

    private final TrainerService trainerService;

    // =========================================================================
    // 2. REGISTRATION (Public)
    // =========================================================================

    @PostMapping
    @Operation(summary = "Register new trainer", description = "Public endpoint - no authentication required")
    @ApiResponse(responseCode = "201", description = "Trainer created successfully",
            content = @Content(schema = @Schema(implementation = UserCreateResponse.class)))
    @ApiResponse(responseCode = "400", description = "Validation error",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    public ResponseEntity<UserCreateResponse> registerTrainer(
            @Valid @RequestBody TrainerCreateRequest request
    ) {
        log.info("Registering new trainer: {} {}", request.firstName(), request.lastName());

        UserCreateResponse response = trainerService.createTrainer(request);

        return ResponseEntity.status(HttpStatus.CREATED)
                .body(response);
    }

    // =========================================================================
    // 8. GET PROFILE
    // =========================================================================

    @GetMapping("/{username}")
    @Operation(summary = "Get trainer profile", description = "Requires authentication - user can only access own profile")
    @SecurityRequirement(name = "bearerAuth")
    @ApiResponse(responseCode = "200", description = "Trainer profile",
            content = @Content(schema = @Schema(implementation = TrainerWithListView.class)))
    @ApiResponse(responseCode = "401", description = "Missing or invalid JWT token",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    @ApiResponse(responseCode = "403", description = "Access denied",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    @ApiResponse(responseCode = "404", description = "Trainer not found",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    @SelfService
    public ResponseEntity<TrainerWithListView> getTrainerProfile(
            @Parameter(description = "Trainer username") @PathVariable String username
    ) {
        log.info("Getting profile for trainer '{}'", username);

        TrainerWithListView response = trainerService.getTrainer(username);
        return ResponseEntity.ok(response);
    }

    // =========================================================================
    // 9. UPDATE PROFILE
    // =========================================================================

    @PutMapping("/{username}")
    @Operation(summary = "Update trainer profile")
    @SecurityRequirement(name = "bearerAuth")
    @ApiResponse(responseCode = "200", description = "Updated trainer profile",
            content = @Content(schema = @Schema(implementation = TrainerWithListView.class)))
    @ApiResponse(responseCode = "400", description = "Validation error",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    @ApiResponse(responseCode = "401", description = "Missing or invalid JWT token",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    @ApiResponse(responseCode = "403", description = "Access denied",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    @ApiResponse(responseCode = "404", description = "Trainer not found",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    @SelfService
    public ResponseEntity<TrainerWithListView> updateTrainerProfile(
            @Parameter(description = "Trainer username") @PathVariable String username,
            @Valid @RequestBody TrainerUpdateRequest request
    ) {
        log.info("Updating profile for trainer '{}'", username);

        TrainerWithListView updated = trainerService.updateTrainer(request);
        return ResponseEntity.ok(updated);
    }

    // =========================================================================
    // 4. CHANGE PASSWORD
    // =========================================================================

    @PutMapping("/{username}/password")
    @Operation(summary = "Change trainer password")
    @SecurityRequirement(name = "bearerAuth")
    @ApiResponse(responseCode = "200", description = "Password changed successfully")
    @ApiResponse(responseCode = "400", description = "Validation error",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    @ApiResponse(responseCode = "401", description = "Missing or invalid JWT token",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    @ApiResponse(responseCode = "403", description = "Access denied",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    @SelfService
    public ResponseEntity<Void> changePassword(
            @Parameter(description = "Trainer username") @PathVariable String username,
            @Valid @RequestBody PasswordChangeRequest request
    ) {
        log.info("Changing password for trainer '{}'", username);

        trainerService.changePassword(request);
        return ResponseEntity.ok().build();
    }

    // =========================================================================
    // 16. ACTIVATE/DEACTIVATE (Non-idempotent)
    // =========================================================================

    @PatchMapping("/{username}/status")
    @Operation(summary = "Activate or deactivate trainer",
            description = "Non-idempotent operation. Returns 409 if already in requested state.")
    @SecurityRequirement(name = "bearerAuth")
    @ApiResponse(responseCode = "200", description = "Status updated successfully")
    @ApiResponse(responseCode = "401", description = "Missing or invalid JWT token",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    @ApiResponse(responseCode = "403", description = "Access denied",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    @ApiResponse(responseCode = "409", description = "User already in requested state",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    @SelfService
    public ResponseEntity<Void> updateActivationStatus(
            @Parameter(description = "Trainer username") @PathVariable String username,
            @Valid @RequestBody ActivationRequest request
    ) {
        log.info("Updating activation status for trainer '{}' to {}", username, request.active());

        // Check current status for non-idempotent behavior
        TrainerWithListView current = trainerService.getTrainer(username);

        if (request.active() && current.active()) {
            throw new UserAlreadyActiveException(username);
        }
        if (!request.active() && !current.active()) {
            throw new UserAlreadyInactiveException(username);
        }

        if (request.active()) {
            trainerService.activate(request);
        } else {
            trainerService.deactivate(request);
        }

        return ResponseEntity.ok().build();
    }

    // =========================================================================
    // 13. GET TRAININGS
    // =========================================================================

    @GetMapping("/{username}/trainings")
    @Operation(summary = "Get trainer's trainings", description = "Returns training sessions with optional filters")
    @SecurityRequirement(name = "bearerAuth")
    @ApiResponse(responseCode = "200", description = "List of trainings",
            content = @Content(array = @ArraySchema(schema = @Schema(implementation = TrainingView.class))))
    @ApiResponse(responseCode = "401", description = "Missing or invalid JWT token",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    @ApiResponse(responseCode = "403", description = "Access denied",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
    @Timed(
            value = "gym.http.trainer.training.search",
            description = "Training search for Trainer endpoint",
            percentiles = {0.95, 0.99},
            histogram = true
    )
    @SelfService
    public ResponseEntity<List<TrainingView>> getTrainings(
            @Parameter(description = "Trainer username")
            @PathVariable String username,
            @Parameter(description = "Filter: start date (inclusive)")
            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate fromDate,
            @Parameter(description = "Filter: end date (inclusive)")
            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate toDate,
            @Parameter(description = "Filter: trainee name")
            @RequestParam(required = false) String traineeName
    ) {
        log.info("Getting trainings for trainer '{}'", username);

        TrainingSearchRequestForTrainer searchRequest = new TrainingSearchRequestForTrainer(
                username,
                fromDate,
                toDate,
                traineeName
        );

        List<TrainingView> response = trainerService.getTrainings(searchRequest);
        return ResponseEntity.ok(response);
    }
}