# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Gym CRM API
  description: |
    REST API for Gym Customer Relationship Management system.
    
    ## Authentication
    
    This API uses **JWT (JSON Web Token)** Bearer authentication for protected endpoints.
    
    ### How to authenticate:
    1. Register as a **Trainee** (`POST /api/v1/trainees`) or **Trainer** (`POST /api/v1/trainers`)
    2. Login with your credentials (`POST /api/v1/auth/login`)
    3. Copy the `token` from the response
    4. Click the **Authorize** button (🔓) above
    5. Enter: `Bearer <your-token>` (or just the token - Swagger UI adds "Bearer" prefix)
    6. Click **Authorize** to save
    
    ### Public endpoints (no authentication required):
    - `POST /api/v1/auth/login` - User login
    - `POST /api/v1/trainees` - Register trainee
    - `POST /api/v1/trainers` - Register trainer  
    - `GET /api/v1/training-types` - List training types
    - `GET /health` - Health check
    
    ### Token details:
    - **Format**: JWT (JSON Web Token)
    - **Expiration**: 24 hours
    - **Claims**: `sub` (username), `role` (TRAINEE/TRAINER), `iat`, `exp`
  version: 1.0.0
  contact:
    name: Ilker Sahin
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/trainers:
    post:
      tags: [Trainers]
      summary: Register new trainer
      description: |
        Create a new trainer account.
        
        **Public endpoint - no authentication required.**
        
        Returns auto-generated username and password. Save these credentials to login later.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainerCreateRequest'
            example:
              firstName: Jane
              lastName: Smith
              specialization: Fitness
      responses:
        '201':
          description: Trainer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreateResponse'
              example:
                username: Jane.Smith
                password: xY7zAb3dEf
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/trainers/{username}:
    get:
      tags: [Trainers]
      summary: Get trainer profile
      description: |
        Retrieve trainer profile information.
        
        **🔒 Requires authentication** - User can only access their own profile.
      parameters:
        - $ref: '#/components/parameters/username'
      responses:
        '200':
          description: Trainer profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainerWithListView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Trainers]
      summary: Update trainer profile
      description: |
        Update trainer profile information.
        
        **🔒 Requires authentication** - User can only update their own profile.
        
        Note: `specialization` cannot be changed after registration.
      parameters:
        - $ref: '#/components/parameters/username'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainerUpdateRequest'
      responses:
        '200':
          description: Updated trainer profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainerWithListView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/trainers/{username}/password:
    put:
      tags: [Trainers]
      summary: Change trainer password
      description: |
        Change trainer's password.
        
        **🔒 Requires authentication** - User can only change their own password.
      parameters:
        - $ref: '#/components/parameters/username'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
      responses:
        '200':
          description: Password changed successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/trainers/{username}/status:
    patch:
      tags: [Trainers]
      summary: Activate or deactivate trainer
      description: |
        Change trainer's active status.
        
        **🔒 Requires authentication** - User can only change their own status.
        
        **Non-idempotent operation**: Returns `409 Conflict` if user is already in the requested state.
      parameters:
        - $ref: '#/components/parameters/username'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivationRequest'
      responses:
        '200':
          description: Status updated successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: User already in requested state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/trainers/{username}/trainings:
    get:
      tags: [Trainers]
      summary: Get trainer's training sessions
      description: |
        Returns training sessions conducted by the trainer with optional filters.
        
        **🔒 Requires authentication** - User can only access their own trainings.
      parameters:
        - $ref: '#/components/parameters/username'
        - name: fromDate
          in: query
          description: Filter trainings from this date (inclusive)
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          description: Filter trainings until this date (inclusive)
          schema:
            type: string
            format: date
        - name: traineeName
          in: query
          description: Filter by trainee username
          schema:
            type: string
      responses:
        '200':
          description: List of trainings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainingView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==========================================================================
  # TRAININGS
  # ==========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication.
        
        **How to obtain a token:**
        1. Login via `POST /api/v1/auth/login`
        2. Copy the `token` from the response
        3. Click "Authorize" and enter the token
        
        **Token format:** `eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
        
        **Token expiration:** 24 hours
        
        **Token claims:**
        - `sub`: Username
        - `role`: User role (TRAINEE or TRAINER)
        - `iat`: Issued at timestamp
        - `exp`: Expiration timestamp
  parameters:
    username:
      name: username
      in: path
      required: true
      schema:
        type: string
      description: |
        User's unique username (format: FirstName.LastName)
      example: John.Doe
  responses:
    BadRequest:
      description: Validation error - request body contains invalid data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-01-15T10:30:00"
            status: 400
            error: Bad Request
            message: Validation failed
            path: /api/v1/trainees
    Unauthorized:
      description: |
        Missing or invalid JWT token.
        
        **Possible causes:**
        - No `Authorization` header provided
        - Token doesn't start with `Bearer `
        - Token is expired
        - Token signature is invalid
        - Token has been invalidated (user logged out)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-01-15T10:30:00"
            status: 401
            error: Unauthorized
            message: Missing or invalid Authorization header
            path: /api/v1/trainees/John.Doe
    Forbidden:
      description: |
        Access denied - user cannot access this resource.
        
        Users can only access their own profile and data.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-01-15T10:30:00"
            status: 403
            error: Forbidden
            message: "User 'John.Doe' cannot access resources of 'Jane.Smith'"
            path: /api/v1/trainees/Jane.Smith
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-01-15T10:30:00"
            status: 404
            error: Not Found
            message: Trainee not found with username 'Unknown.User'
            path: /api/v1/trainees/Unknown.User
  schemas:
...
    TrainerCreateRequest:
      type: object
      required: [firstName, lastName, specialization]
      properties:
        firstName:
          type: string
          description: Trainer's first name
          minLength: 1
          example: Jane
        lastName:
          type: string
          description: Trainer's last name
          minLength: 1
          example: Smith
        specialization:
          type: string
          description: Training specialization type
          enum: [Fitness, Yoga, Zumba, Stretching, Resistance]
          example: Fitness
    TrainerUpdateRequest:
      type: object
      required: [username, firstName, lastName, active]
      properties:
        username:
          type: string
          description: Username (must match path parameter)
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        active:
          type: boolean
          description: Account active status
        specialization:
          type: string
          description: "**Read-only** - ignored on update"
    TrainerWithListView:
      type: object
      properties:
        username:
          type: string
          example: Jane.Smith
        firstName:
          type: string
          example: Jane
        lastName:
          type: string
          example: Smith
        active:
          type: boolean
          example: true
        specialization:
          type: string
          example: Fitness
        trainees:
          type: array
          description: List of assigned trainees
          items:
            $ref: '#/components/schemas/UserInfo'
...
    TrainingView:
      type: object
      properties:
        trainingName:
          type: string
          example: Morning Cardio Session
        trainingDate:
          type: string
          format: date
          example: "2024-01-20"
        durationMinutes:
          type: integer
          example: 60
        trainingType:
          type: string
          example: Fitness
        traineeUsername:
          type: string
          example: John.Doe
        trainerUsername:
          type: string
          example: Jane.Smith
...
    UserCreateResponse:
      type: object
      description: Response after successful user registration
      properties:
        username:
          type: string
          description: Auto-generated username (FirstName.LastName format)
          example: John.Doe
        password:
          type: string
          description: Auto-generated password - **save this, it won't be shown again!**
          example: aB3dEf9xYz
    PasswordChangeRequest:
      type: object
      required: [username, oldPassword, newPassword]
      properties:
        username:
          type: string
          description: Username (must match path parameter)
        oldPassword:
          type: string
          format: password
          description: Current password
        newPassword:
          type: string
          format: password
          description: New password
          minLength: 1
    ActivationRequest:
      type: object
      required: [username, active]
      properties:
        username:
          type: string
          description: Username (must match path parameter)
        active:
          type: boolean
          description: Desired active status (true = activate, false = deactivate)
    UserInfo:
      type: object
      properties:
        username:
          type: string
          example: John.Doe
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
    ErrorResponse:
      type: object
      description: Standard error response format
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
          example: "2024-01-15T10:30:00"
        status:
          type: integer
          description: HTTP status code
          example: 401
        error:
          type: string
          description: HTTP status text
          example: Unauthorized
        message:
          type: string
          description: Detailed error message
          example: Invalid username or password
        path:
          type: string
          description: Request path that caused the error
          example: /api/v1/auth/login